feat(DiscountRate): 할인율을 Integer로 변경

+++ src/main/java/coupon/coupon/domain/DiscountRate.java
@@ -3,7 +3,7 @@ package coupon.coupon.domain;
 import jakarta.persistence.Embeddable;
 
 @Embeddable
-public record DiscountRate(Double discountRate) {
+public record DiscountRate(Integer discountRate) {
 
     private static final double MINIMUM_DISCOUNT_RATE = 3;
     private static final double MAXIMUM_DISCOUNT_RATE = 20;
@@ -12,7 +12,7 @@ public record DiscountRate(Double discountRate) {
         validateRateRange(discountRate);
     }
 
-    public void validateRateRange(final Double discountRate) {
+    public void validateRateRange(final Integer discountRate) {
         if (discountRate < MINIMUM_DISCOUNT_RATE || discountRate > MAXIMUM_DISCOUNT_RATE) {
             throw new IllegalArgumentException("할인율은 3% 이상 20% 이하의 값이어야 합니다.");
         }
